1. moved mixpanel util functions to a separate file for reuse in track.ts 2. added mixpanel user identification to user log-in flow 3. implemented TrackButton on trial-expired CTAs 4. refactored track.ts to use re-usable functions
