+++ apps/app/components/welcome/featured/interstitial.tsx
@@ -18,6 +18,7 @@ import LoggedOutComponent from "@/components/modals/logged-out";
 import { TrialExpiredModal } from "@/components/modals/trial-expired-modal";
 import InterstitialDecor from "./interstitial-decor";
 import track from "@/lib/track";
+import { TrackedButton } from "@/components/analytics/TrackedButton";
 
 interface ExamplePrompt {
   prompt: string;
@@ -482,7 +483,11 @@ const Interstitial: React.FC<InterstitialProps> = ({
               </div>
 
               <div className="flex gap-4 mt-8">
-                <Button
+                <TrackedButton
+                  trackingEvent="daydream_prompt_selected_continue"
+                  trackingProperties={{
+                    prompt: selectedPrompt,
+                  }}
                   onClick={handlePromptContinue}
                   disabled={!selectedPrompt || isPromptLoading}
                   size="lg"
@@ -495,7 +500,7 @@ const Interstitial: React.FC<InterstitialProps> = ({
                   ) : (
                     "Continue"
                   )}
-                </Button>
+                </TrackedButton>
               </div>
               <TermsNotice />
             </div>
